pyproject.toml
[project]
name = "ai-pi-generator"
version = "0.1.0"
description = "Auto-PI Research Scaffold"
authors = [{name = "AutoPI", email = "auto@pi.org"}]
dependencies = [
    "arxiv>=2.4.0",
    "geopandas>=1.1.2",
    "langchain-anthropic>=1.3.4",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pandas>=2.3.3",
    "pyalex>=0.21",
    "pyarrow>=23.0.1",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scholarly>=1.7.11",
    "semanticscholar>=0.11.0",
    "tenacity>=9.1.4",
]
requires-python = "==3.10.*"

[tool.pytest.ini_options]
markers = [
    "live_openalex: marks tests that call the real OpenAlex API (require OPENALEX_API_KEY, OPENALEX_EMAIL, pyalex)",
    "live_llm: marks tests that call real LLM APIs (require GEMINI_API_KEY / GOOGLE_API_KEY)",
]
